# test-scenarios/tc03-nonbreaking/openapi.yaml
openapi: "3.0.0"

info:
  title: SpecShield Test API
  description: Sample User Management API for testing SpecShield GitHub Integration
  version: "1.1.0"
  contact:
    name: SpecShield
    url: https://specshield.io

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /users:
    get:
      summary: Get all users
      operationId: listUsers
      tags: [Users]
      responses:
        "200":
          description: List of users returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create a new user
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid request body

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          description: Unique user ID
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found

    delete:
      summary: Delete a user
      operationId: deleteUser
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found

  # ── NEW ENDPOINT ADDED (non-breaking addition) ──────────────────────────
  /products:
    get:
      summary: Get all products
      operationId: listProducts
      tags: [Products]
      parameters:
        - name: category
          in: query
          required: false
          description: Filter by product category
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      summary: Get product by ID
      operationId: getProductById
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "404":
          description: Product not found

components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"
        createdAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"

    # ── NEW SCHEMA (non-breaking addition) ──────────────────────────────
    Product:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: integer
          example: 101
        name:
          type: string
          example: "Laptop Pro"
        price:
          type: number
          format: double
          example: 1299.99
        category:
          type: string
          example: "Electronics"
        inStock:
          type: boolean
          example: true